Moved annotation helpers out of PostGresQueryBuilder to better follow SOLID principles
