+++ src/main/java/com/revature/util/querinator/PostgresQueryBuilder.java
@@ -3,6 +3,7 @@ package com.revature.util.querinator;
 import com.revature.exceptions.AnnotationNotFound;
 import com.revature.exceptions.InvalidInput;
 import com.revature.util.annotation.*;
+import com.revature.util.querinator.annotationhelper.AnnotationGetters;
 import com.revature.util.querinator.crud.CreateBasedQueries;
 import com.revature.util.querinator.crud.DeleteBasedQueries;
 import com.revature.util.querinator.crud.ReadBasedQueries;
@@ -54,17 +55,19 @@ public class PostgresQueryBuilder<T> {
         // Ensures the pojo is supposed to be persisted to a table
         if (!obj.getClass().isAnnotationPresent(Entity.class)) { throw new AnnotationNotFound("Entity annotation not found!!"); }
 
+        AnnotationGetters annoGetter = new AnnotationGetters();
+
         // Holds the table name related to our POJO
-        String tableName = getTableName(obj);
+        String tableName = annoGetter.getTableName(obj);
 
         // Get the queries column names
-        ArrayDeque<String> queryColumns = getColumnNames(obj);
+        ArrayDeque<String> queryColumns = annoGetter.getColumnNames(obj);
 
         // Get the queries values
-        ArrayDeque<Object> queryValues = getValues(obj);
+        ArrayDeque<Object> queryValues = annoGetter.getValues(obj);
 
         // Primary key info [0] will be the pk column name [1] will be the key itself
-        Object[] pkInfo = getPrimaryKey(obj);
+        Object[] pkInfo = annoGetter.getPrimaryKey(obj);
 
         /*
             Will be needed for login functions within switch
@@ -114,7 +117,7 @@ public class PostgresQueryBuilder<T> {
 
                 readGenerator = new ReadBasedQueries(conn);
 
-                loginInfo = getLoginInfoByUsername(obj);
+                loginInfo = annoGetter.getLoginInfoByUsername(obj);
 
                 query = readGenerator.buildLoginByUsername(tableName, loginInfo);
 
@@ -124,7 +127,7 @@ public class PostgresQueryBuilder<T> {
 
                 readGenerator = new ReadBasedQueries(conn);
 
-                loginInfo = getLoginInfoByEmail(obj);
+                loginInfo = annoGetter.getLoginInfoByEmail(obj);
 
                 query = readGenerator.buildLoginByEmail(tableName, loginInfo);
 
@@ -141,312 +144,6 @@ public class PostgresQueryBuilder<T> {
         }
 
         return query;
-
-    }
-
-    /**
-     *
-     * @param obj
-     * @return tableName
-     *
-     */
-    private String getTableName(Object obj) {
-
-        // Get the objects class
-        Class clazz = obj.getClass();
-
-        // Return value
-        String tableName = "";
-
-        // POJO Class level annotations
-        Annotation[] classAnnotations = clazz.getAnnotations();
-
-        // Loop through our class level annotations
-        for(Annotation ano : classAnnotations) {
-
-            // If there is a table annotation...
-            if (ano instanceof Table) {
-                Table table = (Table) clazz.getAnnotation(ano.annotationType());
-
-                // Set our table name variable to the annotation table_name value
-                tableName = table.table_name();
-
-            }
-
-        }
-
-        return tableName;
-    }
-
-    /**
-     *
-     * @param obj
-     * @return ArrayDeque of column names
-     */
-    private ArrayDeque<String> getColumnNames(Object obj) {
-
-        // Get the objects class
-        Class clazz = obj.getClass();
-
-        // POJO Class's fields
-        Field[] classFields = clazz.getDeclaredFields();
-
-        // Return value
-        ArrayDeque<String> queryColumns = new ArrayDeque<>();
-
-        // Field level annotations
-        Annotation[] fieldAnno;
-
-        // Start looping through the class fields
-        for(Field field : classFields) {
-
-            // Grab the current fields annotations
-            fieldAnno = field.getAnnotations();
-
-            // Loop through the annotations in the previous step
-            for (Annotation ano : fieldAnno) {
-
-                // Check for the Column annotation
-                if (ano instanceof Column) {
-
-                    Column column = (Column) field.getAnnotation(ano.annotationType());
-
-                    // If so add the column name to our column deque
-                    queryColumns.add(column.name());
-
-                }
-            }
-        }
-
-        return queryColumns;
-
-    }
-
-    /**
-     *
-     * @param obj
-     * @return ArrayDeque of mixed type values
-     * @throws IllegalAccessException
-     */
-    private ArrayDeque<Object> getValues(Object obj) throws IllegalAccessException {
-
-        // Get the objects class
-        Class clazz = obj.getClass();
-
-        // POJO Class's fields
-        Field[] classFields = clazz.getDeclaredFields();
-
-        // This will hold out values part of the query
-        ArrayDeque fieldHolder = new ArrayDeque();
-
-        // Field level annotations
-        Annotation[] fieldAnno;
-
-        // Start looping through the class fields
-        for(Field field : classFields) {
-
-            // Grab the current fields annotations
-            fieldAnno = field.getAnnotations();
-
-            // Loop through the annotations in the previous step
-            for (Annotation ano : fieldAnno) {
-
-                // Check for the Column annotation
-                if (ano instanceof Column) {
-
-                    // Allow this script to grab the private fields
-                    field.setAccessible(true);
-
-                    //...Otherwise just add it to the ArrayDeque as is
-                    fieldHolder.add(field.get(obj));
-
-                    // Set the private field back to inaccessible
-                    field.setAccessible(false);
-                }
-            }
-        }
-
-        return fieldHolder;
-
-    }
-
-    /**
-     *
-     * @param obj
-     * @return int (primary key)
-     * @throws IllegalAccessException
-     */
-    private Object[] getPrimaryKey(Object obj) throws IllegalAccessException, InvalidInput {
-
-        // Get the objects class
-        Class clazz = obj.getClass();
-
-        // POJO Class's fields
-        Field[] classFields = clazz.getDeclaredFields();
-
-        // Annotation holder
-        Annotation[] fieldAnno;
-
-        // Return values
-        int primaryKey = -1;
-        String primaryColumnName = "";
-        Object[] returnArray = new Object[2];
-
-        for (Field field : classFields) {
-
-            // Grab the current fields annotations
-            fieldAnno = field.getAnnotations();
-
-            // Loop through the annotations in the previous step
-            for (Annotation ano : fieldAnno) {
-
-                // Check for the Column annotation
-                if (ano instanceof Primary) {
-
-                    Primary pkAnno = (Primary) field.getAnnotation(ano.annotationType());
-
-                    primaryColumnName = pkAnno.name();
-
-                    // Allow this script to grab the private fields
-                    field.setAccessible(true);
-
-                    if (field.get(obj) != null) {
-
-                        primaryKey = (int) field.get(obj);
-
-                    } else { throw new InvalidInput("Primary key cannot be null!"); }
-
-                    // Stop others from grabbing the private field
-                    field.setAccessible(false);
-
-                }
-            }
-        }
-
-        if (primaryKey == -1) { throw new InvalidInput("Primary key is non-existent!"); }
-
-        returnArray[0] = primaryColumnName;
-        returnArray[1] = primaryKey;
-
-
-        return returnArray;
-
-    }
-
-    public Object[][] getLoginInfoByUsername(Object obj) throws IllegalAccessException {
-
-        Object[][] returnArray = new String[3][2];
-
-        // Get the objects class
-        Class clazz = obj.getClass();
-
-        // POJO Class's fields
-        Field[] classFields = clazz.getDeclaredFields();
-
-        // Field level annotations
-        Annotation[] fieldAnno;
-
-        // Start looping through the class fields
-        for(Field field : classFields) {
-
-            // Grab the current fields annotations
-            fieldAnno = field.getAnnotations();
-
-            // Loop through the annotations in the previous step
-            for (Annotation ano : fieldAnno) {
-
-                // Check for the Username annotation
-                if (ano instanceof Username) {
-
-                    // Allow this script to grab the private fields
-                    field.setAccessible(true);
-
-
-                    ///...And add to the return array
-                    returnArray[0][0] = field.getAnnotation(Column.class).name();
-                    returnArray[0][1] = field.get(obj);
-
-
-                    // Set the private field back to inaccessible
-                    field.setAccessible(false);
-                }
-
-                // Check for the Username annotation
-                if (ano instanceof Password) {
-
-                    // Allow this script to grab the private fields
-                    field.setAccessible(true);
-
-
-                    // add to the return array
-                    returnArray[1][0] = field.getAnnotation(Column.class).name();
-                    returnArray[1][1] = field.get(obj);
-
-
-                    // Set the private field back to inaccessible
-                    field.setAccessible(false);
-                }
-            }
-        }
-
-        return returnArray;
-
-    }
-
-    public Object[][] getLoginInfoByEmail(Object obj) throws IllegalAccessException {
-
-        Object[][] returnArray = new String[3][2];
-
-        // Get the objects class
-        Class clazz = obj.getClass();
-
-        // POJO Class's fields
-        Field[] classFields = clazz.getDeclaredFields();
-
-        // Field level annotations
-        Annotation[] fieldAnno;
-
-        // Start looping through the class fields
-        for(Field field : classFields) {
-
-            // Grab the current fields annotations
-            fieldAnno = field.getAnnotations();
-
-            // Loop through the annotations in the previous step
-            for (Annotation ano : fieldAnno) {
-
-                // Check for the Username annotation
-                if (ano instanceof Email) {
-
-                    // Allow this script to grab the private fields
-                    field.setAccessible(true);
-
-                    // Add to the return array
-                    returnArray[0][0] = field.getAnnotation(Column.class).name();
-                    returnArray[0][1] = field.get(obj);
-
-
-                    // Set the private field back to inaccessible
-                    field.setAccessible(false);
-                }
-
-                // Check for the Username annotation
-                if (ano instanceof Password) {
-
-                    // Allow this script to grab the private fields
-                    field.setAccessible(true);
-
-                    // Add to the return array
-                    returnArray[1][0] = field.getAnnotation(Column.class).name();
-                    returnArray[1][1] = field.get(obj);
-
-                    // Set the private field back to inaccessible
-                    field.setAccessible(false);
-                }
-            }
-        }
-
-        return returnArray;
 
     }
